\name{deadliestDayPerYear}
\alias{deadliestDayPerYear}
\title{Deadliest Day per Year}
\description{Takes as input a year, from 2014 to 2022, and outputs deadliest day of the week.}
\usage{
deadliestDayPerYear(x)
}
\arguments{
year
}
\value{
An text representation of deadliest day of the week for the given year.
}
\examples{
deadliestDayPerYear(2015)
}