\name{totalDeathsPerYear}
\alias{totalDeathsPerYear}
\title{Total Deaths of Year}
\description{Takes as input a year, from 2014 to 2022, and outputs the number of fatalities.}
\usage{
totalDeathsPerYear(x)
}
\arguments{
year
}
\value{
An integer representation of the number of fatalities for the given year.
}
\examples{
totalDeathsPerYear(2015)
}