\name{totalInjuriesPerYear}
\alias{totalInjuriesPerYear}
\title{Total Injuries of the Year}
\description{Takes as input a year, from 2014 to 2022, and outputs the number of injuries.}
\usage{
totalInjuriesPerYear(x)
}
\arguments{
year
}
\value{
An integer representation of the number of injuries for the given year.
}
\examples{
totalInjuriesPerYear(2015)
}